requirements.txt
email-validator>=2.2.0
flask>=3.0.3
openai>=1.54.3
python-dotenv>=1.0.0
rich>=13.7.0
tiktoken>=0.6.0
python-markdown>=3.5.2
gunicorn>=21.2.0 